refactor: move repository depdencies to provider
