+++ internal/provider/publiccloud/repository/shared/error.go
