+++ internal/provider/publiccloud/repository/shared/error_response_test.go
